requirements.txt
Flask>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
googlesearch-python>=1.3.0
lxml>=4.9.0